+++ tests/TestCase.php
@@ -24,6 +24,17 @@ class TestCase extends \PHPUnit\Framework\TestCase
                 'port' => 22,
                 'root' => '/path/to/application',
                 'branch' => 'master',
+                'scripts' => [
+                    'before-deploy' => [
+                        '@php @composer {{ @release }}/artisan some-command'
+                    ]
+                ],
+                'php' => [
+                    'bin' => 'php-7.4',
+                ],
+                'composer' => [
+                    'bin' => 'composer.phar',
+                ]
             ], [
                 'name' => 'server-2',
             ],